feat: refactor kafka source API and enhance virtual sources resolving logic (#37)
- added support for binary format for key and value. This means, that
kafka message key or value is read as is, without casting to other
types. Thus, it is up to user to use virtual sources/streams
functionality to cast column to a desired type suitable for data quality
checks.
- added boolean flag to enable schema ID subtraction from kafka value:
when schema registry is used, the schema ID is embedded into kafka value
(magic byte + 4 bytes of schema ID). Therefore, in order to parse value,
first it is required to subtract schema ID from it.
- changed Avro schema API by adding boolean flag that enables or disable
default values checks.
- refactor logic for resolving virtual sources: now they are resolved
not in order but rather with respect to their parents dependencies.
- added tests to verify virtual source resolving logic.
- documentation updated with respect to API changes.
- configuration api version is updated to 1.5
- fixes: 
    - updates SQLIte dependency version: security patch [VDB-248999]. 
    - fixed JoinVirtualSourceReader by adding aliases to dataframes that are
being joined. This is needed to avoid ambiguous column referencing in
the resultant dataframe.